shopthing/requirements.txt
amqp==5.0.6
asgiref==3.4.1
billiard==3.6.4.0
cached-property==1.5.2
celery==5.1.2
certifi==2021.5.30
charset-normalizer==2.0.6
click==7.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Django==3.2.7
django-classy-tags==2.0.0
django-cms==3.9.0
django-formtools==2.3
django-sekizai==2.0.0
django-treebeard==4.5.1
djangocms-admin-style==2.0.2
gunicorn
idna==3.2
idpay==0.0.1
importlib-metadata==4.8.1
isodate==0.6.0
kavenegar==1.1.2
kombu==5.1.0
Pillow==8.3.2
prompt-toolkit==3.0.20
psycopg2-binary==2.9.1
python-decouple==3.5
pytz==2021.1
redis==3.5.3
requests==2.26.0
six==1.16.0
sqlparse==0.4.1
typing-extensions==3.10.0.2
urllib3==1.26.6
vine==5.0.0
wcwidth==0.2.5
zipp==3.6.0
